requirements.txt
flask==2.0.1
python-owasp-zap-v2.4==0.0.20
requests==2.26.0
python-dotenv==0.19.0
Werkzeug==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
click==8.0.1
urllib3==1.26.6
idna==3.2
certifi==2021.5.30
chardet==4.0.0